+++ source/gray_code.cpp
@@ -9,14 +9,14 @@ using namespace cppcoro;
  * @brief Binary Reflexed Gray Code Generator
  *
  * The function `brgc_gen` is a generator function that generates binary
- * reflexed gray code. It takes an input parameter `n` of type `size_t` and
- * returns a `recursive_generator<size_t>`.
+ * reflexed gray code. It takes an input parameter `n` of type `int` and
+ * returns a `recursive_generator<int>`.
  *
- * @param n The parameter `n` represents the number of bits in the binary reflexed gray code sequence
- * to be generated.
- * @return recursive_generator<size_t>
+ * @param[in] n The parameter `n` represents the number of bits in the binary
+ * reflexed gray code sequence to be generated.
+ * @return recursive_generator<int>
  */
-auto brgc_gen(size_t n) -> recursive_generator<size_t> {
+auto brgc_gen(int n) -> recursive_generator<int> {
     if (n < 1) {
         co_return;
     }
